+++ client/src/main/java/net/luxvacuos/voxel/client/world/chunks/Chunk.java
@@ -102,7 +102,7 @@ public class Chunk {
 	}
 
 	public void update(Dimension dimension, DimensionService service, Camera camera, float delta) {
-		distance = Vector3f.sub(camera.getPosition(), new Vector3f(posX, posY, posZ), null).lengthSquared();
+		distance = (float) Vector3f.sub(camera.getPosition(), new Vector3f(posX, posY, posZ), null).lengthSquared();
 
 		if (!created && !creating) {
 			creating = true;
