+++ utils/src/main/java/net/luxvacuos/igl/vector/ReadableVector2f.java
@@ -38,10 +38,10 @@ public interface ReadableVector2f extends ReadableVector {
 	/**
 	 * @return x
 	 */
-	float getX();
+	double getX();
 
 	/**
 	 * @return y
 	 */
-	float getY();
+	double getY();
 }
