+++ utils/src/main/java/net/luxvacuos/igl/vector/ReadableVector3f.java
@@ -38,5 +38,5 @@ public interface ReadableVector3f extends ReadableVector2f {
 	/**
 	 * @return z
 	 */
-	float getZ();
+	double getZ();
 }
